sel4c/requirements.txt
# DO NOT include azure-functions-worker in this file
# The Python Worker is managed by Azure Functions platform
# Manually managing azure-functions-worker may cause unexpected issues

azure-functions
wordcloud==1.9.2
moviepy==1.0.3
speechrecognition==3.10.0
gensim==4.3.2
nltk==3.8.1
